factor some tests
